bump patch version (again)
I forgot to bump the patch version for the clang-format release awhile ago and now I want to make a new release so that I can re-publish the website with the seminar slides. I don't want to break the last tag which is being held in ldmx-sw right now, so here we are.